+++ src/assets/docs/images/blur-cyan.png
