+++ src/assets/docs/images/blur-indigo.png
