+++ src/components/docs/PrevNextLinks.jsx
@@ -4,7 +4,7 @@ import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import clsx from 'clsx'
 
-import { navigation } from '@/lib/syntax/navigation'
+import { navigation } from '@/lib/docs/navigation'
 
 function ArrowIcon(props) {
   return (
