--- src/styles/landing/landing.css
@@ -1,5 +1,0 @@
-@layer components {
-  .hero-title {
-    @apply bg-linear-to-r from-indigo-600 to-sky-500 bg-clip-text text-transparent;
-  }
-}
